+++ server/controllers/html.js
@@ -1,0 +1,5 @@
+const html = (route) => (req, reply) => {
+  reply.view("runner/htmlViewer");
+};
+
+module.exports = html;
