+++ server/controllers/index.js
@@ -1,0 +1,5 @@
+const common = require("./common");
+const html = require("./html");
+const node = require("./node");
+
+module.exports = { common, html, node };
