+++ server/controllers/node.js
@@ -1,0 +1,5 @@
+const node = (route) => (req, reply) => {
+  reply.view("runner/nodeViewer");
+};
+
+module.exports = node;
